backend/requirements.txt
# EVERYA Backend Dependencies
flask>=2.3.0
flask-cors>=4.0.0
mysql-connector-python>=8.0.0
werkzeug>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0